+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.4.1] - 2026-06-07
+
 ### Added
 
 - **Accumulator type inference via `defineGet`/`definePost`/`defineRoute` helpers.** (#91)